+++ test/common/models/Mailbox.js
@@ -23,7 +23,7 @@ describe('The Mailbox class', function() {
       var mailbox = new jmap.Mailbox({}, 'id', 'name');
 
       expect(mailbox.parentId).to.equal(null);
-      expect(mailbox.role).to.equal(null);
+      expect(mailbox.role).to.equal(jmap.MailboxRole.UNKNOWN);
       expect(mailbox.sortOrder).to.equal(0);
       expect(mailbox.mustBeOnlyMailbox).to.equal(false);
       expect(mailbox.mayReadItems).to.equal(false);
@@ -42,7 +42,7 @@ describe('The Mailbox class', function() {
       var mailbox = new jmap.Mailbox({}, 'id', 'name', {});
 
       expect(mailbox.parentId).to.equal(null);
-      expect(mailbox.role).to.equal(null);
+      expect(mailbox.role).to.equal(jmap.MailboxRole.UNKNOWN);
       expect(mailbox.sortOrder).to.equal(0);
       expect(mailbox.mustBeOnlyMailbox).to.equal(false);
       expect(mailbox.mayReadItems).to.equal(false);
@@ -125,7 +125,7 @@ describe('The Mailbox class', function() {
       var mailbox = jmap.Mailbox.fromJSONObject({}, { id: 'id', name: 'name' });
 
       expect(mailbox.parentId).to.equal(null);
-      expect(mailbox.role).to.equal(null);
+      expect(mailbox.role).to.equal(jmap.MailboxRole.UNKNOWN);
       expect(mailbox.sortOrder).to.equal(0);
       expect(mailbox.mustBeOnlyMailbox).to.equal(false);
       expect(mailbox.mayReadItems).to.equal(false);
@@ -163,7 +163,7 @@ describe('The Mailbox class', function() {
       expect(mailbox.id).to.equal('id');
       expect(mailbox.name).to.equal('name');
       expect(mailbox.parentId).to.equal('parentId');
-      expect(mailbox.role).to.equal('inbox');
+      expect(mailbox.role).to.equal(jmap.MailboxRole.INBOX);
       expect(mailbox.sortOrder).to.equal(1);
       expect(mailbox.mustBeOnlyMailbox).to.equal(true);
       expect(mailbox.mayReadItems).to.equal(true);